Fix another windows compiler warning

+++ test/integration/twoProcessesPubSub.cc
@@ -86,7 +86,7 @@ void cbVector(const ignition::msgs::Vector3d &/*_msg*/)
 }
 
 //////////////////////////////////////////////////
-void cbRaw(const char * /*_msgData*/, const int /*_size*/,
+void cbRaw(const char * /*_msgData*/, const size_t /*_size*/,
            const ignition::transport::MessageInfo &/*_info*/)
 {
   cbRawExecuted = true;
